@@ -22,7 +22,7 @@
     "chalk": "^2.4.1",
     "chromedriver": "^2.43.1",
     "config": "^2.0.1",
-    "copy-webpack-plugin": "^4.5.4",
+    "copy-webpack-plugin": "^4.6.0",
     "cross-spawn": "^6.0.5",
     "css-loader": "^1.0.1",
     "dotenv": "^6.1.0",
@@ -36,7 +36,7 @@
     "faker": "^4.1.0",
     "file-loader": "^2.0.0",
     "git-describe": "^4.0.3",
-    "graphql-cli": "^2.16.7",
+    "graphql-cli": "^2.17.0",
     "handlebars": "^4.0.12",
     "handlebars-loader": "^1.7.0",
     "html-loader": "^0.5.5",
@@ -53,31 +53,34 @@
     "wdio": "^3.0.0",
     "wdio-chromedriver-service": "^0.1.1",
     "wdio-dot-reporter": "^0.0.10",
-    "wdio-mocha-framework": "^0.6.3",
+    "wdio-mocha-framework": "^0.6.4",
     "wdio-spec-reporter": "^0.1.5",
     "webdriverio": "^4.14.0",
-    "webpack": "^4.23.1",
+    "webpack": "^4.25.1",
     "webpack-cleanup-plugin": "^0.5.1",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.10",
     "webpack-merge": "^4.1.4"
   },
   "dependencies": {
+    "@blueprintjs/core": "^3.7.0",
     "bootstrap": "^3.3.7",
     "change-case": "^3.0.1",
+    "classnames": "^2.2.6",
     "core-js": "^2.5.7",
     "create-react-class": "^15.6.3",
     "d3": "^5.7.0",
     "deep-equal": "^1.0.1",
-    "email-addresses": "^3.0.2",
     "draft-convert": "^2.1.4",
     "draft-js": "^0.10.5",
+    "email-addresses": "^3.0.2",
     "file-saver": "^1.3.3",
     "graphiql": "^0.12.0",
     "graphql": "^0.13.2",
     "hopscotch": "^0.3.1",
     "imports-loader": "^0.8.0",
     "leaflet": "^1.3.4",
+    "leaflet-gesture-handling": "^1.1.8",
     "locale-compare-polyfill": "^0.0.2",
     "lodash": "^4.17.11",
     "moment": "^2.22.2",
@@ -90,7 +93,9 @@
     "react-autosuggest-ie11-compatible": "https://github.com/VassilIordanov/react-autosuggest-ie11-compatible/releases/download/v9.4.2/react-autosuggest-ie11-compatible.tar.gz",
     "react-bootstrap": "^0.32.4",
     "react-confirm-bootstrap": "^5.3.0",
+    "react-container-dimensions": "^1.4.1",
     "react-dom": "^16.6.0",
+    "react-mosaic-component": "^2.0.0",
     "react-redux": "^5.1.0",
     "react-redux-loading-bar": "^4.1.0",
     "react-router": "^4.3.1",
