@@ -1,125 +1,175 @@
+import React from 'react'
 import PropTypes from 'prop-types'
-import React, {Component} from 'react'
-import API from 'api'
-import Settings from 'Settings'
 import autobind from 'autobind-decorator'
-import {Button} from 'react-bootstrap'
 
+import {Popover, Overlay} from 'react-bootstrap'
 import BarChart from 'components/BarChart'
-import Fieldset from 'components/Fieldset'
-import ReportCollection from 'components/ReportCollection'
-
-import {Report} from 'models'
-
-import _isEqual from 'lodash/isEqual'
+import ReportCollection, {FORMAT_MAP, FORMAT_SUMMARY, FORMAT_TABLE} from 'components/ReportCollection'
 
 import { connect } from 'react-redux'
 import LoaderHOC, {mapDispatchToProps} from 'HOC/LoaderHOC'
+import Settings from 'Settings'
 
-const d3 = require('d3')
-const chartId = 'not_approved_reports_chart'
-const GQL_CHART_FIELDS =  /* GraphQL */`
-  uuid
-  advisorOrg { uuid, shortName }
-`
+import ReportsVisualisation, {propTypes as rvPropTypes} from 'components/ReportsVisualisation'
+import ContainerDimensions from 'react-container-dimensions'
+import { IconNames } from '@blueprintjs/icons'
+
 const BarChartWithLoader = connect(null, mapDispatchToProps)(LoaderHOC('isLoading')('data')(BarChart))
+const Context = React.createContext()
 
 /*
  * Component displaying reports submitted for approval up to the given date but
  * which have not been approved yet. They are displayed in different
  * presentation forms: chart, summary, table and map.
  */
-class PendingApprovalReports extends Component {
-  static propTypes = {
-    queryParams: PropTypes.object,
-    showLoading: PropTypes.func.isRequired,
-    hideLoading: PropTypes.func.isRequired,
-  }
+class PendingApprovalReports extends ReportsVisualisation {
+  static propTypes = {...rvPropTypes}
 
   constructor(props) {
     super(props)
 
+    this.advisorOrgLabel = Settings.fields.advisor.org.name
+    this.CHART_ID = 'not_approved_reports_chart'
+    this.GQL_CHART_FIELDS =  /* GraphQL */`
+      uuid
+      advisorOrg { uuid, shortName }
+    `
+    this.VISUALIZATIONS = [
+      {
+        id: 'par-chart',
+        icons: [IconNames.GROUPED_BAR_CHART],
+        title: `Chart by ${this.advisorOrgLabel}`,
+        renderer: this.getBarChart,
+      },
+      {
+        id: 'par-collection',
+        icons: [IconNames.PANEL_TABLE],
+        title: `Reports by ${this.advisorOrgLabel}`,
+        renderer: this.getReportCollection,
+      },
+      {
+        id: 'par-map',
+        icons: [IconNames.MAP],
+        title: `Map by ${this.advisorOrgLabel}`,
+        renderer: this.getReportMap,
+      },
+    ]
+    this.INITIAL_LAYOUT = {
+      direction: 'column',
+      first: {
+        direction: 'row',
+        first: this.VISUALIZATIONS[0].id,
+        second: this.VISUALIZATIONS[1].id,
+      },
+      second: this.VISUALIZATIONS[2].id,
+    }
+    this.DESCRIPTION = `The reports are grouped by ${this.advisorOrgLabel}.
+      In order to see the list of pending approval reports for an organization,
+      click on the bar corresponding to the organization.`
+
     this.state = {
       graphData: [],
-      reports: {list: []},
+      reports: {},
+      allReports: [],
       reportsPageNum: 0,
-      focusedOrg: '',
+      focusedSelection: '',
+      graphPopover: null,
+      hoveredBar: null,
       updateChart: true,  // whether the chart needs to be updated
       isLoading: false
     }
   }
 
-  render() {
-    const focusDetails = this.focusDetails
-    return (
-      <div>
-        <p className="help-text">{`Grouped by ${Settings.fields.advisor.org.name}`}</p>
-        <p className="chart-description">
-          {`The reports are grouped by ${Settings.fields.advisor.org.name}. In
-            order to see the list of pending approval reports for an organization,
-            click on the bar corresponding to the organization.`}
-        </p>
-        <BarChartWithLoader
-          chartId={chartId}
-          data={this.state.graphData}
-          xProp='advisorOrg.uuid'
-          yProp='notApproved'
-          xLabel='advisorOrg.shortName'
-          onBarClick={this.goToOrg}
-          updateChart={this.state.updateChart}
-          isLoading={this.state.isLoading}
-        />
-        <Fieldset
-          title={`Pending Approval Reports ${focusDetails.titleSuffix}`}
-          id='not-approved-reports-details'
-          action={!focusDetails.resetFnc
-            ? '' : <Button onClick={() => this[focusDetails.resetFnc]()}>{focusDetails.resetButtonLabel}</Button>
-          } >
-          {this.state.reports.list && <ReportCollection paginatedReports={this.state.reports} goToPage={this.goToReportsPage} />}
-        </Fieldset>
+  get additionalReportParams() {
+    return {advisorOrgUuid: this.state.focusedSelection.advisorOrg.uuid}
+  }
+
+  @autobind
+  getBarChart(id) {
+    return <Context.Consumer>{context => (
+      <div className="non-scrollable">
+        <ContainerDimensions>{({width, height}) => (
+          <BarChartWithLoader
+            width={width}
+            height={height}
+            chartId={this.chartId}
+            data={context.graphData}
+            xProp='advisorOrg.uuid'
+            yProp='notApproved'
+            xLabel='advisorOrg.shortName'
+            onBarClick={this.goToSelection}
+            showPopover={this.showPopover}
+            hidePopover={this.hidePopover}
+            updateChart={context.updateChart}
+            selectedBarClass={this.selectedBarClass}
+            selectedBar={context.focusedSelection ? 'bar_' + context.focusedSelection.advisorOrg.uuid : ''}
+            isLoading={context.isLoading}
+          />
+        )}</ContainerDimensions>
+
+        <Overlay
+          show={!!context.graphPopover}
+          placement="top"
+          container={document.body}
+          animation={false}
+          target={() => context.graphPopover}
+        >
+          <Popover id="graph-popover" title={context.hoveredBar && context.hoveredBar.advisorOrg.shortName}>
+            <p style={{textAlign: 'center'}}>{context.hoveredBar && context.hoveredBar.notApproved}</p>
+          </Popover>
+        </Overlay>
       </div>
+    )}</Context.Consumer>
+  }
+
+  @autobind
+  getReportCollection(id)
+  {
+    return <Context.Consumer>{context => (
+      <div className="scrollable">
+        <ReportCollection
+          paginatedReports={context.reports}
+          goToPage={this.goToReportsPage}
+          viewFormats={[FORMAT_TABLE, FORMAT_SUMMARY]}
+        />
+      </div>
+    )}</Context.Consumer>
+  }
+
+  @autobind
+  getReportMap(id)
+  {
+    return <Context.Consumer>{context => (
+      <div className="non-scrollable">
+        <ContainerDimensions>{({width, height}) => (
+          <ReportCollection
+            width={width}
+            height={height}
+            marginBottom={0}
+            reports={context.allReports}
+            viewFormats={[FORMAT_MAP]}
+          />
+        )}</ContainerDimensions>
+      </div>
+    )}</Context.Consumer>
+  }
+
+  render() {
+    return (
+      <Context.Provider value={this.state}>
+        {super.render()}
+      </Context.Provider>
     )
   }
 
-  get focusDetails() {
-    let titleSuffix = ''
-    let resetFnc = ''
-    let resetButtonLabel = ''
-    if (this.state.focusedOrg) {
-      titleSuffix = `for ${this.state.focusedOrg.shortName}`
-      resetFnc = 'goToOrg'
-      resetButtonLabel = 'All organizations'
-    }
-    return {
-      titleSuffix: titleSuffix,
-      resetFnc: resetFnc,
-      resetButtonLabel: resetButtonLabel
-    }
-  }
-
-  fetchData() {
-    this.setState( {isLoading: true} )
-    this.props.showLoading()
-    const pinned_ORGs = Settings.pinned_ORGs
-    const chartQueryParams = {}
-    Object.assign(chartQueryParams, this.props.queryParams)
-    Object.assign(chartQueryParams, {
-      pageNum: 0,
-      pageSize: 0,  // retrieve all the filtered reports
-    })
-    // Query used by the chart
-    const chartQuery = API.query(/* GraphQL */`
-        reportList(query:$chartQueryParams) {
-          totalCount, list {
-            ${GQL_CHART_FIELDS}
-          }
-        }
-      `, {chartQueryParams}, '($chartQueryParams: ReportSearchQueryInput)')
-    const noAdvisorOrg = {
-      uuid: "-1",
-      shortName: `No ${Settings.fields.advisor.org.name}`
-    }
-    Promise.all([chartQuery]).then(values => {
+  @autobind
+  fetchChartData(chartQuery) {
+    return Promise.all([chartQuery]).then(values => {
+      const pinned_ORGs = Settings.pinned_ORGs
+      const noAdvisorOrg = {
+        uuid: "-1",
+        shortName: `No ${this.advisorOrgLabel}`
+      }
       let reportsList = values[0].reportList.list || []
       reportsList = reportsList
         .map(d => { if (!d.advisorOrg) d.advisorOrg = noAdvisorOrg; return d })
@@ -140,68 +190,13 @@ class PendingApprovalReports extends Component {
             }
           })
       })
-      this.props.hideLoading()
-    })
-    this.fetchOrgData()
-  }
-
-  fetchOrgData() {
-    const reportsQueryParams = {}
-    Object.assign(reportsQueryParams, this.props.queryParams)
-    Object.assign(reportsQueryParams, {
-      pageNum: this.state.reportsPageNum,
-      pageSize: 10
-    })
-    if (this.state.focusedOrg) {
-      Object.assign(reportsQueryParams, {advisorOrgUuid: this.state.focusedOrg.uuid})
-    }
-    // Query used by the reports collection
-    let reportsQuery = API.query(/* GraphQL */`
-        reportList(query:$reportsQueryParams) {
-          pageNum, pageSize, totalCount, list {
-            ${ReportCollection.GQL_REPORT_FIELDS}
-          }
-        }
-      `, {reportsQueryParams}, '($reportsQueryParams: ReportSearchQueryInput)')
-    Promise.all([reportsQuery]).then(values => {
-      this.setState({
-        updateChart: false,  // only update the report list
-        reports: values[0].reportList
-      })
     })
   }
 
   @autobind
-  goToReportsPage(newPage) {
-    this.setState({updateChart: false, reportsPageNum: newPage}, () => this.fetchOrgData())
+  updateHighlight(focusedSelection, clear) {
+    super.updateHighlight(focusedSelection ? focusedSelection.advisorOrg.uuid : '', clear)
   }
-
-  resetChartSelection() {
-    d3.selectAll('#' + chartId + ' rect').attr('class', '')
-  }
-
-  @autobind
-  goToOrg(item) {
-    // Note: we set updateChart to false as we do not want to rerender the chart
-    // when changing the focus organization.
-    this.setState({updateChart: false, reportsPageNum: 0, focusedOrg: (item ? item.advisorOrg : '')}, () => this.fetchOrgData())
-    // remove highlighting of the bars
-    this.resetChartSelection()
-    if (item) {
-      // highlight the bar corresponding to the selected organization
-      d3.select('#' + chartId + ' #bar_' + item.advisorOrg.uuid).attr('class', 'selected-bar')
-    }
-  }
-
-  componentDidUpdate(prevProps, prevState) {
-    if (!_isEqual(prevProps.queryParams, this.props.queryParams)) {
-      this.setState({
-        reportsPageNum: 0,
-        focusedOrg: ''  // reset focus when changing the queryParams
-      }, () => this.fetchData())
-    }
-  }
-
 }
 
 export default connect(null, mapDispatchToProps)(PendingApprovalReports)
