@@ -1,5 +1,6 @@
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
+import _isEmpty from 'lodash/isEmpty'
 import { showLoading, hideLoading } from 'react-redux-loading-bar'
 import './LoaderHOC.css'
 
@@ -11,15 +12,6 @@ export const mapDispatchToProps = (dispatch, ownProps) => ({
 const LoaderHOC = (isLoading) => (dataPropName) => (WrappedComponent) => {
     return class LoaderHOC extends Component {
 
-        isEmpty(prop) {
-            return (
-                prop === null ||
-                prop === undefined ||
-                (prop.hasOwnProperty('length') && prop.length === 0) ||
-                (prop.constructor === Object && Object.keys(prop).length === 0)
-            )
-        }
-
         isLoadingData(prop) {
             return (
                 prop ||
@@ -28,18 +20,16 @@ const LoaderHOC = (isLoading) => (dataPropName) => (WrappedComponent) => {
         }
 
         render() {
-            const dataIsEmpty = this.isEmpty(this.props[dataPropName])
+            const dataIsEmpty = _isEmpty(this.props[dataPropName])
             const showLoader =  dataIsEmpty && this.isLoadingData(this.props[isLoading])
 
             if (showLoader) {
                 return <div className='loader'></div>
-            } else {
-                if (dataIsEmpty) {
-                    return <div></div>
-                }
-                else {
-                    return <WrappedComponent {...this.props} />
-                }
+            } else if (dataIsEmpty) {
+                return <div><em>No data</em></div>
+            }
+            else {
+                return <WrappedComponent {...this.props} />
             }
         }
     }
