@@ -23,6 +23,8 @@ function getPropValue(obj, prop) {
 
 class BarChart extends Component {
   static propTypes = {
+    width: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+    height: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
     chartId: PropTypes.string,
     data: PropTypes.array,
     xProp: PropTypes.string.isRequired,
@@ -30,11 +32,18 @@ class BarChart extends Component {
     xLabel: PropTypes.string,
     barClass: PropTypes.string,
     onBarClick: PropTypes.func,
+    showPopover: PropTypes.func,
+    hidePopover: PropTypes.func,
+    selectedBarClass: PropTypes.string,
+    selectedBar: PropTypes.string,
     updateChart: PropTypes.bool
   }
 
   static defaultProps = {
-    barClass: 'bar',
+    width: '100%',
+    barClass: 'bars-group',
+    selectedBarClass: 'selected-bar',
+    selectedBar: '',
     updateChart: true
   }
 
@@ -51,8 +60,15 @@ class BarChart extends Component {
     this.createBarChart()
   }
 
+  isNumeric(value) {
+    return typeof value === 'number'
+  }
+
   createBarChart() {
-    const MARGIN = {top: 20, right: 20}  // left and bottom margins are dynamic
+    const MARGIN = {
+      top: 20, right: 20,
+      left: 40, bottom: 0,  // left and bottom MARGINs are dynamic, these are extra margins
+    }
     let chartData = this.props.data
     let xProp = this.props.xProp  // data property to use for the x-axis domain
     let yProp = this.props.yProp  // data property to use for the y-axis domain
@@ -62,8 +78,9 @@ class BarChart extends Component {
 
     let xScale = d3.scaleBand()
       .domain(chartData.map(function(d) { xLabels[getPropValue(d, xProp)] = getPropValue(d, xLabel); return getPropValue(d, xProp) }))
+    let yMax = d3.max(chartData, function(d) { return getPropValue(d, yProp) })
     let yScale = d3.scaleLinear()
-      .domain([0, d3.max(chartData, function(d) { return getPropValue(d, yProp) })])
+      .domain([0, yMax])
 
     // Calculate the maximum width of the axis labels
     let maxXLabelWidth = 0
@@ -87,7 +104,6 @@ class BarChart extends Component {
       tmpSVG.selectAll('.get_max_width_y_label')
         .data(chartData)
         .enter().append('text')
-        .attr('class', 'y-axis')
         .text(yText)
         .each(yLabelWidth)
         .remove()
@@ -97,14 +113,14 @@ class BarChart extends Component {
     // The left margin depends on the width of the y-axis labels.
     // We add extra margin to make sure that if the label is different because
     // of the automatic formatting the labels are still displayed on the chart.
-    let marginLeft = maxYLabelWidth + 50
+    let marginLeft = maxYLabelWidth + MARGIN.left
     // The bottom margin depends on the width of the x-axis labels.
-    let marginBottom = maxXLabelWidth
+    let marginBottom = maxXLabelWidth + MARGIN.bottom
 
     let chart = d3.select(this.node)
     let chartBox = this.node.getBoundingClientRect()
-    let chartWidth = chartBox.right - chartBox.left
-    let chartHeight = 0.7 * chartWidth
+    let chartWidth = this.isNumeric(this.props.width) ? this.props.width : (chartBox.right - chartBox.left)
+    let chartHeight = this.isNumeric(this.props.height) ? this.props.height : (0.7 * chartWidth)
     let xWidth = chartWidth - marginLeft - MARGIN.right
     let yHeight = chartHeight - MARGIN.top - marginBottom
 
@@ -114,7 +130,10 @@ class BarChart extends Component {
 
     let xAxis = d3.axisBottom(xScale)
       .tickFormat(function(d) { return xLabels[d] })
+
+    let yTicks = Math.min(yMax, 10)
     let yAxis = d3.axisLeft(yScale)
+      .ticks(yTicks, 'd')
 
     chart.selectAll('*').remove()
     chart = chart
@@ -134,17 +153,21 @@ class BarChart extends Component {
     chart.append('g')
       .call(yAxis)
 
-    let bar = chart.selectAll('.bar')
+    const selectedBar = this.props.selectedBar
+    let bar = chart.selectAll(`.${this.props.barClass}`)
       .data(chartData)
       .enter()
       .append('g')
-      .classed('bars-group', true)
+      .classed(this.props.barClass, true)
       .append('rect')
-      .attr('id', function(d, i) { return 'bar_' + getPropValue(d, xProp) })
+      .attr('id', function(d, i) { return `bar_${getPropValue(d, xProp)}` })
+      .classed(this.props.selectedBarClass, function(d, i) { return this.id === selectedBar })
       .attr('x', function(d) { return xScale(getPropValue(d, xProp)) })
       .attr('y', function(d) { return yScale(getPropValue(d, yProp)) })
       .attr('width', xScale.bandwidth())
       .attr('height', function(d) { return yHeight - yScale(getPropValue(d, yProp)) })
+      .on('mouseenter', d => this.props.showPopover && this.props.showPopover(d3.event.target, d))
+      .on('mouseleave', d => this.props.hidePopover && this.props.hidePopover())
     if (onBarClick) {
       bar.on('click', function(d) {
         onBarClick(d)
@@ -153,14 +176,16 @@ class BarChart extends Component {
   }
 
   render() {
-    return <svg id={this.props.chartId} ref={node => this.node = node} width="100%"></svg>
+    return <svg id={this.props.chartId} ref={node => this.node = node} width={this.props.width} height={this.props.height}></svg>
   }
 
   shouldComponentUpdate(nextProps, nextState) {
     // Make sure the chart is only re-rendered if the state or properties have
     // changed. This because we do not want to re-render the chart only in order
     // to highlight a bar in the chart.
-    if (nextProps && !nextProps.updateChart) {
+    if (nextProps && !nextProps.updateChart
+        && nextProps.width === this.props.width
+        && nextProps.height === this.props.height) {
       return false
     }
     return true
