@@ -1,210 +1,206 @@
+import React from 'react'
 import PropTypes from 'prop-types'
-import React, {Component} from 'react'
-import API from 'api'
 import autobind from 'autobind-decorator'
-import {Button} from 'react-bootstrap'
 
+import {Popover, Overlay} from 'react-bootstrap'
 import BarChart from 'components/BarChart'
-import Fieldset from 'components/Fieldset'
-import ReportCollection from 'components/ReportCollection'
-
-import {Report} from 'models'
-
-import _isEqual from 'lodash/isEqual'
+import ReportCollection, {FORMAT_MAP, FORMAT_SUMMARY, FORMAT_TABLE} from 'components/ReportCollection'
 
 import { connect } from 'react-redux'
 import LoaderHOC, {mapDispatchToProps} from 'HOC/LoaderHOC'
 import Settings from 'Settings'
 
+import ReportsVisualisation, {propTypes as rvPropTypes} from 'components/ReportsVisualisation'
+import ContainerDimensions from 'react-container-dimensions'
+import { IconNames } from '@blueprintjs/icons'
+
 import pluralize from 'pluralize'
 
-const d3 = require('d3')
-const chartByTaskId = 'reports_by_task'
-const GQL_CHART_FIELDS =  /* GraphQL */`
-  uuid
-  tasks { uuid, shortName }
-`
 const BarChartWithLoader = connect(null, mapDispatchToProps)(LoaderHOC('isLoading')('data')(BarChart))
+const Context = React.createContext()
 
 /*
  * Component displaying a chart with number of reports per Task.
  */
-class ReportsByTask extends Component {
-  static propTypes = {
-    queryParams: PropTypes.object,
-    showLoading: PropTypes.func.isRequired,
-    hideLoading: PropTypes.func.isRequired,
-  }
+class ReportsByTask extends ReportsVisualisation {
+  static propTypes = {...rvPropTypes}
 
   constructor(props) {
     super(props)
 
+    this.taskShortLabel = Settings.fields.task.shortLabel
+    this.CHART_ID = 'reports_by_task'
+    this.GQL_CHART_FIELDS =  /* GraphQL */`
+      uuid
+      tasks { uuid shortName }
+    `
+    this.VISUALIZATIONS = [
+      {
+        id: 'rbt-chart',
+        icons: [IconNames.GROUPED_BAR_CHART],
+        title: `Chart by ${this.taskShortLabel}`,
+        renderer: this.getBarChart,
+      },
+      {
+        id: 'rbt-collection',
+        icons: [IconNames.PANEL_TABLE],
+        title: `Reports by ${this.taskShortLabel}`,
+        renderer: this.getReportCollection,
+      },
+      {
+        id: 'rbt-map',
+        icons: [IconNames.MAP],
+        title: `Map by ${this.taskShortLabel}`,
+        renderer: this.getReportMap,
+      },
+    ]
+    this.INITIAL_LAYOUT = {
+      direction: 'column',
+      first: {
+        direction: 'row',
+        first: this.VISUALIZATIONS[0].id,
+        second: this.VISUALIZATIONS[1].id,
+      },
+      second: this.VISUALIZATIONS[2].id,
+    }
+    this.DESCRIPTION = `The reports are grouped by ${this.taskShortLabel}.
+      In order to see the list of published reports for a ${this.taskShortLabel},
+      click on the bar corresponding to the ${this.taskShortLabel}.`
+
     this.state = {
-      graphDataByTask: [],
+      graphData: [],
+      reports: {},
+      allReports: [],
       reportsPageNum: 0,
-      focusedTask: '',
+      focusedSelection: '',
+      graphPopover: null,
+      hoveredBar: null,
       updateChart: true,  // whether the chart needs to be updated
       isLoading: false
     }
   }
 
-  render() {
-    const focusDetails = this.getFocusDetails()
-    const taskShortLabel = Settings.fields.task.shortLabel
-    return (
-      <div>
-        <p className="help-text">{`Grouped by ${taskShortLabel}`}</p>
-        <p className="chart-description">
-          {`The reports are grouped by ${taskShortLabel}. In order to see the
-            list of published reports for a ${taskShortLabel}, click on the bar
-            corresponding to the ${taskShortLabel}.`}
-        </p>
-        <BarChartWithLoader
-          chartId={chartByTaskId}
-          data={this.state.graphDataByTask}
-          xProp='task.uuid'
-          yProp='reportsCount'
-          xLabel='task.shortName'
-          onBarClick={this.goToTask}
-          updateChart={this.state.updateChart}
-          isLoading={this.state.isLoading}
-        />
-        <Fieldset
-          title={`Reports by ${taskShortLabel} ${focusDetails.titleSuffix}`}
-          id='cancelled-reports-details'
-          action={!focusDetails.resetFnc
-            ? '' : <Button onClick={() => this[focusDetails.resetFnc]()}>{focusDetails.resetButtonLabel}</Button>
-          } >
-          <ReportCollection paginatedReports={this.state.reports} goToPage={this.goToReportsPage} />
-        </Fieldset>
+  get additionalReportParams() {
+    return {taskUuid: this.state.focusedSelection.task.uuid}
+  }
+
+  @autobind
+  getBarChart(id) {
+    return <Context.Consumer>{context => (
+      <div className="non-scrollable">
+        <ContainerDimensions>{({width, height}) => (
+          <BarChartWithLoader
+            width={width}
+            height={height}
+            chartId={this.chartId}
+            data={context.graphData}
+            xProp='task.uuid'
+            yProp='reportsCount'
+            xLabel='task.shortName'
+            onBarClick={this.goToSelection}
+            showPopover={this.showPopover}
+            hidePopover={this.hidePopover}
+            updateChart={context.updateChart}
+            selectedBarClass={this.selectedBarClass}
+            selectedBar={context.focusedSelection ? 'bar_' + context.focusedSelection.task.uuid : ''}
+            isLoading={context.isLoading}
+          />
+        )}</ContainerDimensions>
+
+        <Overlay
+          show={!!context.graphPopover}
+          placement="top"
+          container={document.body}
+          animation={false}
+          target={() => context.graphPopover}
+        >
+          <Popover id="graph-popover" title={context.hoveredBar && context.hoveredBar.task.shortName}>
+            <p style={{textAlign: 'center'}}>{context.hoveredBar && context.hoveredBar.reportsCount}</p>
+          </Popover>
+        </Overlay>
       </div>
+    )}</Context.Consumer>
+  }
+
+  @autobind
+  getReportCollection(id)
+  {
+    return <Context.Consumer>{context => (
+      <div className="scrollable">
+        <ReportCollection
+          paginatedReports={context.reports}
+          goToPage={this.goToReportsPage}
+          viewFormats={[FORMAT_TABLE, FORMAT_SUMMARY]}
+        />
+      </div>
+    )}</Context.Consumer>
+  }
+
+  @autobind
+  getReportMap(id)
+  {
+    return <Context.Consumer>{context => (
+      <div className="non-scrollable">
+        <ContainerDimensions>{({width, height}) => (
+          <ReportCollection
+            width={width}
+            height={height}
+            marginBottom={0}
+            reports={context.allReports}
+            viewFormats={[FORMAT_MAP]}
+          />
+        )}</ContainerDimensions>
+      </div>
+    )}</Context.Consumer>
+  }
+
+  render() {
+    return (
+      <Context.Provider value={this.state}>
+        {super.render()}
+      </Context.Provider>
     )
   }
 
-  getFocusDetails() {
-    let titleSuffix = ''
-    let resetFnc = ''
-    let resetButtonLabel = ''
-    const allTasks = `All ${pluralize(Settings.fields.task.shortLabel)}`
-
-    if (this.state.focusedTask) {
-      titleSuffix = `for ${this.state.focusedTask.shortName}`
-      resetFnc = 'goToTask'
-      resetButtonLabel = allTasks
-    }
-    return {
-      titleSuffix: titleSuffix,
-      resetFnc: resetFnc,
-      resetButtonLabel: resetButtonLabel
-    }
-  }
-
-  fetchData() {
-    this.setState( {isLoading: true} )
-    this.props.showLoading()
-    const chartQueryParams = {}
-    Object.assign(chartQueryParams, this.props.queryParams)
-    Object.assign(chartQueryParams, {
-      pageNum: 0,
-      pageSize: 0,  // retrieve all the filtered reports
-    })
-    // Query used by the chart
-    const chartQuery = API.query(/* GraphQL */`
-        reportList(query:$chartQueryParams) {
-          totalCount, list {
-            ${GQL_CHART_FIELDS}
-          }
-        }
-      `, {chartQueryParams}, '($chartQueryParams: ReportSearchQueryInput)')
-    const noTaskMessage = `No ${Settings.fields.task.shortLabel}`
-    const noTask = {
-      uuid: "-1",
-      shortName: noTaskMessage,
-      longName: noTaskMessage
-    }
-    Promise.all([chartQuery]).then(values => {
-      let simplifiedValues = values[0].reportList.list ? values[0].reportList.list.map(d => {return {reportUuid: d.uuid, tasks: d.tasks.map(p => p.uuid)}}): []
-      let tasks = values[0].reportList.list ? values[0].reportList.list.map(d => d.tasks) : []
-      tasks = [].concat.apply([], tasks)
-        .filter((item, index, d) => d.findIndex(t => {return t.uuid === item.uuid }) === index)
-        .sort((a, b) => a.shortName.localeCompare(b.shortName))
-      // add No Task item, in order to relate to reports without Tasks
-      tasks.push(noTask)
+  @autobind
+  fetchChartData(chartQuery) {
+    return Promise.all([chartQuery]).then(values => {
+      const noTaskMessage = `No ${Settings.fields.task.shortLabel}`
+      const noTask = {
+        uuid: "-1",
+        shortName: noTaskMessage,
+        longName: noTaskMessage
+      }
+      let graphData = []
+      let reportsList = values[0].reportList.list || []
+      if (!!reportsList.length) {
+        let simplifiedValues = reportsList.map(d => {return {reportUuid: d.uuid, tasks: d.tasks.map(p => p.uuid)}})
+        let tasks = reportsList.map(d => d.tasks)
+        tasks = [].concat.apply([], tasks)
+          .filter((item, index, d) => d.findIndex(t => {return t.uuid === item.uuid }) === index)
+          .sort((a, b) => a.shortName.localeCompare(b.shortName))
+        // add No Task item, in order to relate to reports without Tasks
+        tasks.push(noTask)
+        graphData = tasks.map(d => {
+          let r = {}
+          r.task = d
+          r.reportsCount = (d.uuid === noTask.uuid)
+            ? simplifiedValues.filter(item => item.tasks.length === 0).length
+            : simplifiedValues.filter(item => item.tasks.indexOf(d.uuid) > -1).length
+          return r
+        })
+      }
       this.setState({
         isLoading: false,
         updateChart: true,  // update chart after fetching the data
-        graphDataByTask: tasks
-          .map(d => {
-            let r = {}
-            r.task = d
-            r.reportsCount = (d.uuid ? simplifiedValues.filter(item => item.tasks.indexOf(d.uuid) > -1).length : simplifiedValues.filter(item => item.tasks.length === 0).length)
-            return r}),
-      })
-      this.props.hideLoading()
-    })
-    this.fetchTaskData()
-  }
-
-  fetchTaskData() {
-    const reportsQueryParams = {}
-    Object.assign(reportsQueryParams, this.props.queryParams)
-    Object.assign(reportsQueryParams, {
-      pageNum: this.state.reportsPageNum,
-      pageSize: 10
-    })
-    if (this.state.focusedTask) {
-      Object.assign(reportsQueryParams, {taskUuid: this.state.focusedTask.uuid})
-    }
-    // Query used by the reports collection
-    const reportsQuery = API.query(/* GraphQL */`
-        reportList(query:$reportsQueryParams) {
-          pageNum, pageSize, totalCount, list {
-            ${ReportCollection.GQL_REPORT_FIELDS}
-          }
-        }
-      `, {reportsQueryParams}, '($reportsQueryParams: ReportSearchQueryInput)')
-    Promise.all([reportsQuery]).then(values => {
-      this.setState({
-        updateChart: false,  // only update the report list
-        reports: values[0].reportList
+        graphData: graphData
       })
     })
   }
 
   @autobind
-  goToReportsPage(newPage) {
-    this.setState({updateChart: false, reportsPageNum: newPage}, () => this.fetchTaskData())
-  }
-
-  resetChartSelection(chartId) {
-    d3.selectAll('#' + chartId + ' rect').attr('class', '')
-  }
-
-  @autobind
-  goToTask(item) {
-    // Note: we set updateChart to false as we do not want to re-render the chart
-    // when changing the focus task.
-    this.setState({updateChart: false, reportsPageNum: 0, focusedTask: (item ? item.task : '')}, () => this.fetchTaskData())
-    // remove highlighting of the bars
-    this.resetChartSelection(chartByTaskId)
-    if (item) {
-      // highlight the bar corresponding to the selected task
-      d3.select('#' + chartByTaskId + ' #bar_' + item.task.uuid).attr('class', 'selected-bar')
-    }
-  }
-
-
-  componentDidMount() {
-    this.fetchData()
-  }
-
-  componentDidUpdate(prevProps, prevState) {
-    if (!_isEqual(prevProps.queryParams, this.props.queryParams)) {
-      this.setState({
-        reportsPageNum: 0,
-        focusedTask: ''  // reset focus when changing the queryParams
-      }, () => this.fetchData())
-    }
+  updateHighlight(focusedSelection, clear) {
+    super.updateHighlight(focusedSelection ? focusedSelection.task.uuid : '', clear)
   }
 }
 
