@@ -9,7 +9,6 @@ import autobind from 'autobind-decorator'
 import {Person, Report} from 'models'
 
 import AppContext from 'components/AppContext'
-import { DEFAULT_PAGE_PROPS, CLEAR_SEARCH_PROPS } from 'actions'
 import { connect } from 'react-redux'
 
 class BaseMyReports extends Page {
@@ -20,7 +19,7 @@ class BaseMyReports extends Page {
 	}
 
 	constructor(props) {
-		super(props, DEFAULT_PAGE_PROPS, CLEAR_SEARCH_PROPS)
+		super(props)
 
 		this.state = {
 			draft: null,
