@@ -8,7 +8,8 @@ import pluralize from 'pluralize'
 
 import Fieldset from 'components/Fieldset'
 import Breadcrumbs from 'components/Breadcrumbs'
-import ReportCollection from 'components/ReportCollection'
+import DailyRollupChart from 'components/DailyRollupChart'
+import ReportCollection, {FORMAT_MAP, FORMAT_SUMMARY, FORMAT_TABLE, GQL_REPORT_FIELDS} from 'components/ReportCollection'
 import CalendarButton from 'components/CalendarButton'
 import ButtonToggleGroup from 'components/ButtonToggleGroup'
 import Form from 'components/Form'
@@ -21,11 +22,16 @@ import utils from 'utils'
 import API from 'api'
 
 import AppContext from 'components/AppContext'
-import { DEFAULT_PAGE_PROPS, CLEAR_SEARCH_PROPS } from 'actions'
 import { withRouter } from 'react-router-dom'
 import { connect } from 'react-redux'
+import LoaderHOC, {mapDispatchToProps as loaderMapDispatchToProps} from 'HOC/LoaderHOC'
 
-var d3 = null/* required later */
+import MosaicLayout from 'components/MosaicLayout'
+import ContainerDimensions from 'react-container-dimensions'
+import { IconNames } from '@blueprintjs/icons'
+
+const BarChartWithLoader = connect(null, loaderMapDispatchToProps)(LoaderHOC('isLoading')('data')(DailyRollupChart))
+const Context = React.createContext()
 
 const barColors = {
 	cancelled: '#EC971F',
@@ -56,7 +62,49 @@ class BaseRollupShow extends Page {
 	get rollupEnd() { return moment(this.state.date).endOf('day').hour(18) } // 6:59:59pm today.
 
 	constructor(props) {
-		super(props, DEFAULT_PAGE_PROPS, CLEAR_SEARCH_PROPS)
+		super(props)
+
+		this.CHART_ID = 'reports_by_day_of_week'
+		this.GQL_CHART_FIELDS = /* GraphQL */`
+			org {uuid shortName}
+			released
+			cancelled
+		`
+		this.GQL_MAP_FIELDS = /* GraphQL */`
+			uuid
+			intent
+			location { uuid name lat lng }
+		`
+		this.VISUALIZATIONS = [
+			{
+				id: 'rbdow-chart',
+				icons: [IconNames.GROUPED_BAR_CHART],
+				title: 'Chart by organization',
+				renderer: this.getBarChart,
+			},
+			{
+				id: 'rbdow-collection',
+				icons: [IconNames.PANEL_TABLE],
+				title: 'Reports by organization',
+				renderer: this.getReportCollection,
+			},
+			{
+				id: 'rbdow-map',
+				icons: [IconNames.MAP],
+				title: 'Map by organization',
+				renderer: this.getReportMap,
+			},
+		]
+		this.INITIAL_LAYOUT = {
+			direction: 'row',
+			first: this.VISUALIZATIONS[0].id,
+			second: {
+				direction: 'column',
+				first: this.VISUALIZATIONS[1].id,
+				second: this.VISUALIZATIONS[2].id,
+			}
+		}
+		this.DESCRIPTION = `Number of reports released today per organization.`
 
 		const qs = utils.parseQueryString(props.location.search)
 		this.state = {
@@ -69,6 +117,8 @@ class BaseRollupShow extends Page {
 			maxReportAge: null,
 			hoveredBar: {org: {}},
 			orgType: Organization.TYPE.ADVISOR_ORG,
+			updateChart: true,  // whether the chart needs to be updated
+			isLoading: false
 		}
 		this.previewPlaceholderUrl = API.addAuthParams("/help")
 	}
@@ -92,22 +142,6 @@ class BaseRollupShow extends Page {
 		if (!this.state.date.isSame(prevState.date, 'day') || prevState.maxReportAge !== this.state.maxReportAge) {
 			this.loadData()
 		}
-		else {
-			this.renderGraph()
-		}
-	}
-
-	componentDidMount() {
-		super.componentDidMount()
-
-		if (d3) {
-			return
-		}
-
-		import('d3').then(importedModule => {
-			d3 = importedModule
-			this.forceUpdate()
-		})
 	}
 
 	fetchData(props) {
@@ -116,71 +150,124 @@ class BaseRollupShow extends Page {
 			return
 		}
 
-		const rollupReportsQuery = {
+		this.setState({isLoading: true})
+		this.props.showLoading()
+		Promise.all([
+			// Query used by the chart
+			this.fetchChartData(this.runChartQuery(...this.chartQueryParams())),
+			this.fetchReportData(true)
+		]).then(() => this.props.hideLoading())
+	}
+
+	fetchReportData(includeAll) {
+		// Query used by the reports collection
+		const queries = [this.runReportsQuery(this.reportsQueryParams(false), false)]
+		if (includeAll) {
+			// Query used by the map
+			queries.push(this.runReportsQuery(this.reportsQueryParams(true), true))
+		}
+		return Promise.all(queries).then(values => {
+			const stateUpdate = {
+				updateChart: false,  // only update the report list
+				reports: values[0].reportList
+			}
+			if (includeAll) {
+				Object.assign(stateUpdate, {
+					allReports: values[1].reportList.list
+				})
+			}
+			this.setState(stateUpdate)
+		})
+	}
+
+	chartQueryParams = () => {
+		let chartQuery = 'rollupGraph(startDate: $startDate, endDate: $endDate'
+		let chartQueryParamsDef = '($startDate: Long!, $endDate: Long!'
+		const chartQueryParams = {
+			startDate: this.rollupStart.valueOf(),
+			endDate: this.rollupEnd.valueOf(),
+		}
+		if (this.state.focusedOrg) {
+			if (this.state.orgType === Organization.TYPE.PRINCIPAL_ORG) {
+				chartQuery += ', principalOrganizationUuid: $principalOrganizationUuid'
+				chartQueryParamsDef += ', $principalOrganizationUuid: String!'
+				chartQueryParams.principalOrganizationUuid = this.state.focusedOrg.uuid
+			} else {
+				chartQuery += ' ,advisorOrganizationUuid: $advisorOrganizationUuid'
+				chartQueryParamsDef += ', $advisorOrganizationUuid: String!'
+				chartQueryParams.advisorOrganizationUuid = this.state.focusedOrg.uuid
+			}
+		} else if (this.state.orgType) {
+			chartQuery += ', orgType: $orgType'
+			chartQueryParamsDef += ', $orgType: OrganizationType!'
+			chartQueryParams.orgType = this.state.orgType
+		}
+		chartQuery += ')'
+		chartQueryParamsDef += ')'
+		Object.assign(chartQueryParams, {
+			pageNum: 0,
+			pageSize: 0,  // retrieve all the filtered reports
+		})
+		return [chartQuery, chartQueryParams, chartQueryParamsDef]
+	}
+
+	runChartQuery = (chartQuery,chartQueryParams, chartQueryParamsDef) => {
+		return API.query(/* GraphQL */`
+			${chartQuery} {
+				${this.GQL_CHART_FIELDS}
+			}`, chartQueryParams, chartQueryParamsDef)
+	}
+
+	reportsQueryParams = (forMap) => {
+		const reportsQueryParams = {
 			state: [Report.STATE.RELEASED], //Specifically excluding cancelled engagements.
 			releasedAtStart: this.rollupStart.valueOf(),
 			releasedAtEnd: this.rollupEnd.valueOf(),
 			engagementDateStart: moment(this.rollupStart).subtract(this.state.maxReportAge, 'days').valueOf(),
 			sortBy: "ENGAGEMENT_DATE",
 			sortOrder: "DESC",
-			pageNum: this.state.reportsPageNum,
-			pageSize: 10,
 		}
-		Object.assign(rollupReportsQuery, this.getSearchQuery(props))
-
-		let rollupGraphQuery = 'rollupGraph(startDate: $startDate, endDate: $endDate'
-		let rollupGraphVariableDef = '($startDate: Long!, $endDate: Long!'
-		const rollupGraphVariables = {
-			startDate: rollupReportsQuery.releasedAtStart,
-			endDate: rollupReportsQuery.releasedAtEnd,
-		}
+		Object.assign(reportsQueryParams, this.getSearchQuery(this.props))
+		Object.assign(reportsQueryParams, {
+			pageNum: forMap ? 0 : this.state.reportsPageNum,
+			pageSize: forMap ? 0 : 10
+		})
 		if (this.state.focusedOrg) {
 			if (this.state.orgType === Organization.TYPE.PRINCIPAL_ORG) {
-				rollupReportsQuery.principalOrgUuid = this.state.focusedOrg.uuid
-				rollupReportsQuery.includePrincipalOrgChildren = true
-				rollupGraphQuery += ', principalOrganizationUuid: $principalOrganizationUuid'
-				rollupGraphVariableDef += ', $principalOrganizationUuid: String!'
-				rollupGraphVariables.principalOrganizationUuid = this.state.focusedOrg.uuid
+				reportsQueryParams.principalOrgUuid = this.state.focusedOrg.uuid
+				reportsQueryParams.includePrincipalOrgChildren = true
 			} else {
-				rollupReportsQuery.advisorOrgUuid = this.state.focusedOrg.uuid
-				rollupReportsQuery.includeAdvisorOrgChildren = true
-				rollupGraphQuery += ' ,advisorOrganizationUuid: $advisorOrganizationUuid'
-				rollupGraphVariableDef += ', $advisorOrganizationUuid: String!'
-				rollupGraphVariables.advisorOrganizationUuid = this.state.focusedOrg.uuid
+				reportsQueryParams.advisorOrgUuid = this.state.focusedOrg.uuid
+				reportsQueryParams.includeAdvisorOrgChildren = true
 			}
-		} else if (this.state.orgType) {
-			rollupGraphQuery += ', orgType: $orgType'
-			rollupGraphVariableDef += ', $orgType: OrganizationType!'
-			rollupGraphVariables.orgType = this.state.orgType
 		}
-		rollupGraphQuery += ') {org {uuid shortName} released cancelled}'
-		rollupGraphVariableDef += ')'
+		return reportsQueryParams
+	}
 
-		let reportQuery = API.query(/* GraphQL */`
-			reportList(query:$rollupReportsQuery) {
+	runReportsQuery = (reportsQueryParams, forMap) => {
+		return API.query(/* GraphQL */`
+			reportList(query:$reportsQueryParams) {
 				pageNum, pageSize, totalCount, list {
-					${ReportCollection.GQL_REPORT_FIELDS}
+					${forMap ? this.GQL_MAP_FIELDS : GQL_REPORT_FIELDS}
 				}
-			}
-		`, {rollupReportsQuery}, '($rollupReportsQuery: ReportSearchQueryInput)')
+			}`, {reportsQueryParams}, '($reportsQueryParams: ReportSearchQueryInput)')
+	}
 
-		let graphQuery = API.query(/* GraphQL */
-				rollupGraphQuery, rollupGraphVariables, rollupGraphVariableDef
-		)
-
+	@autobind
+	fetchChartData(chartQuery) {
+		return Promise.all([chartQuery]).then(values => {
 		const pinned_ORGs = Settings.pinned_ORGs
-
-		return Promise.all([reportQuery, graphQuery]).then(values => {
 			this.setState({
-				reports: values[0].reportList,
-				graphData: values[1].rollupGraph
+				isLoading: false,
+				updateChart: true,  // update chart after fetching the data
+				graphData: values[0].rollupGraph
 					.map(d => {d.org = d.org || {uuid: "-1", shortName: "Other"}; return d})
 					.sort((a, b) => {
 						let a_index = pinned_ORGs.indexOf(a.org.shortName)
 						let b_index = pinned_ORGs.indexOf(b.org.shortName)
 						if (a_index<0) {
 							let nameOrder = a.org.shortName.localeCompare(b.org.shortName)
-							return (b_index<0) ?  (nameOrder === 0 ? a.org.uuid - b.org.uuid : nameOrder)  : 1
+							return (b_index<0) ? (nameOrder === 0 ? a.org.uuid - b.org.uuid : nameOrder) : 1
 						}
 						else {
 							return (b_index<0) ? -1 : a_index-b_index
@@ -190,61 +277,116 @@ class BaseRollupShow extends Page {
 		})
 	}
 
+	@autobind
+	getBarChart(id) {
+		return <Context.Consumer>{context => (
+			<div className="scrollable-y">
+				<ContainerDimensions>{({width}) => (
+						<BarChartWithLoader
+							width={width}
+							chartId={this.CHART_ID}
+							data={context.graphData}
+							onBarClick={this.goToOrg}
+							showPopover={this.showPopover}
+							hidePopover={this.hidePopover}
+							updateChart={context.updateChart}
+							isLoading={context.isLoading}
+							barColors={barColors}
+						/>
+				)}</ContainerDimensions>
+
+				<Overlay
+					show={!!context.graphPopover}
+					placement="top"
+					container={document.body}
+					animation={false}
+					target={() => context.graphPopover}
+				>
+					<Popover id="graph-popover" title={context.hoveredBar && context.hoveredBar.org.shortName}>
+						<p>Released: {context.hoveredBar && context.hoveredBar.released}</p>
+						<p>Cancelled: {context.hoveredBar && context.hoveredBar.cancelled}</p>
+						<p>Click to view details</p>
+					</Popover>
+				</Overlay>
+
+				<div className="graph-legend">
+					<div style={{...legendCss, background: barColors.verified}}></div> Released reports:&nbsp;
+					<strong>{context.graphData.reduce((acc, org) => acc + org.released, 0)}</strong>
+				</div>
+				<div className="graph-legend">
+					<div style={{...legendCss, background: barColors.cancelled}}></div> Cancelled engagements:&nbsp;
+					<strong>{context.graphData.reduce((acc, org) => acc + org.cancelled, 0)}</strong>
+				</div>
+			</div>
+		)}</Context.Consumer>
+	}
+
+	@autobind
+	getReportCollection(id)
+	{
+		return <Context.Consumer>{context => (
+			<div className="scrollable">
+				<ReportCollection
+					paginatedReports={context.reports}
+					goToPage={this.goToReportsPage}
+					viewFormats={[FORMAT_TABLE, FORMAT_SUMMARY]}
+				/>
+			</div>
+		)}</Context.Consumer>
+	}
+
+	@autobind
+	getReportMap(id)
+	{
+		return <Context.Consumer>{context => (
+			<div className="non-scrollable">
+				<ContainerDimensions>{({width, height}) => (
+					<ReportCollection
+						width={width}
+						height={height}
+						marginBottom={0}
+						reports={context.allReports}
+						viewFormats={[FORMAT_MAP]}
+					/>
+				)}</ContainerDimensions>
+			</div>
+		)}</Context.Consumer>
+	}
+
 	render() {
+		const flexStyle = {display: 'flex', flexDirection: 'column', height: '100%', flex: 1}
+
 		return (
-			<div>
+			<div id="daily-rollup" style={flexStyle}>
 				<Breadcrumbs items={[[`Rollup for ${this.dateStr}`, '/rollup']]} />
 				<Messages error={this.state.error} success={this.state.success} />
 
 				<Fieldset title={
 					<span>
-						Daily Rollup - {this.dateLongStr}
+						Daily Rollup{this.state.focusedOrg && ` for ${this.state.focusedOrg.shortName}`} - {this.dateLongStr}
 						<CalendarButton onChange={this.changeRollupDate} value={this.state.date.toISOString()} style={calendarButtonCss} />
+						{this.state.focusedOrg
+							? <Button onClick={() => this.goToOrg()}>All organizations</Button>
+							: <ButtonToggleGroup value={this.state.orgType} onChange={this.changeOrgType}>
+								<Button value={Organization.TYPE.ADVISOR_ORG}>{pluralize(Settings.fields.advisor.org.name)}</Button>
+								<Button value={Organization.TYPE.PRINCIPAL_ORG}>{pluralize(Settings.fields.principal.org.name)}</Button>
+							  </ButtonToggleGroup>
+						}
 					</span>
 				} action={
-					<div>
+					<span>
 						<Button href={this.previewPlaceholderUrl} target="rollup" onClick={this.printPreview}>Print</Button>
 						<Button onClick={this.toggleEmailModal} bsStyle="primary">Email rollup</Button>
-					</div>
-				}>
-					<p className="help-text">Number of reports released today per organization</p>
-					<svg ref={el => this.graph = el} style={{width: '100%'}} />
-
-					<Overlay
-						show={!!this.state.graphPopover}
-						placement="top"
-						container={document.body}
-						animation={false}
-						target={() => this.state.graphPopover}
-					>
-						<Popover id="graph-popover" title={this.state.hoveredBar.org.shortName}>
-							<p>Released: {this.state.hoveredBar.released}</p>
-							<p>Cancelled: {this.state.hoveredBar.cancelled}</p>
-							<p>Click to view details</p>
-						</Popover>
-					</Overlay>
-
-					<div className="graph-legend">
-						<div style={{...legendCss, background: barColors.verified}}></div> Released reports:&nbsp;
-						<strong>{this.state.graphData.reduce((acc, org) => acc + org.released, 0)}</strong>
-					</div>
-					<div className="graph-legend">
-						<div style={{...legendCss, background: barColors.cancelled}}></div> Cancelled engagements:&nbsp;
-						<strong>{this.state.graphData.reduce((acc, org) => acc + org.cancelled, 0)}</strong>
-					</div>
-				</Fieldset>
-
-				<Fieldset
-					title={`Reports ${this.state.focusedOrg ? `for ${this.state.focusedOrg.shortName}` : ''}`}
-					action={!this.state.focusedOrg
-						? <ButtonToggleGroup value={this.state.orgType} onChange={this.changeOrgType}>
-							<Button value={Organization.TYPE.ADVISOR_ORG}>{pluralize(Settings.fields.advisor.org.name)}</Button>
-							<Button value={Organization.TYPE.PRINCIPAL_ORG}>{pluralize(Settings.fields.principal.org.name)}</Button>
-						</ButtonToggleGroup>
-						: <Button onClick={() => this.goToOrg()}>All organizations</Button>
-					}
-				>
-					<ReportCollection paginatedReports={this.state.reports} goToPage={this.goToReportsPage} />
+					</span>
+				} style={flexStyle}>
+					<Context.Provider value={this.state}>
+						<MosaicLayout
+							style={flexStyle}
+							visualizations={this.VISUALIZATIONS}
+							initialNode={this.INITIAL_LAYOUT}
+							description={this.DESCRIPTION}
+						/>
+					</Context.Provider>
 				</Fieldset>
 
 				{this.renderEmailModal()}
@@ -342,18 +484,18 @@ class BaseRollupShow extends Page {
 	}
 
 	@autobind
-	goToReportsPage(newPageNum) {
-		this.setState({reportsPageNum: newPageNum}, () => this.loadData())
+	goToReportsPage(newPage) {
+		this.setState({updateChart: false, reportsPageNum: newPage}, () => this.fetchReportData(false))
 	}
 
 	@autobind
 	goToOrg(org) {
-		this.setState({reportsPageNum: 0, focusedOrg: org, graphPopover: null}, () => this.loadData())
+		this.setState({reportsPageNum: 0, focusedOrg: org, isLoading: true}, () => this.loadData())
 	}
 
 	@autobind
 	changeOrgType(orgType) {
-		this.setState({orgType}, () => this.loadData())
+		this.setState({orgType, isLoading: true}, () => this.loadData())
 	}
 
 	@autobind
@@ -363,6 +505,16 @@ class BaseRollupShow extends Page {
 			pathname: 'rollup',
 			search: utils.formatQueryString({date: date.valueOf()})
 		})
+	}
+
+	@autobind
+	showPopover(graphPopover, hoveredBar) {
+		this.setState({graphPopover, hoveredBar})
+	}
+
+	@autobind
+	hidePopover() {
+		this.setState({graphPopover: null, hoveredBar: null})
 	}
 
 	@autobind
